Wrap frame retrieval in try/catch
If the player collides with the wild before it completely initializes, the game would crash. Wrapping it in a try block seems like a band-aid but really it is how I want to manage this edge case. In this case I just assign the value to be what it would have been, if the game had not crashed - red.